+++ src/utils/string-utils/index.js
@@ -1,4 +1,10 @@
 // @flow
 export {
-  makeTitleString, normalizeStr, sentenceCase, upperAlphaChars, formatPercent,
+  makeTitleString,
+  normalizeStr,
+  sentenceCase,
+  upperAlphaChars,
+  formatPercent,
+  extractQuotedString,
+  unquoteString,
 } from "./string-util";