+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - localStorage now includes `omni-` prefix.
 - Add support for omni commands. This allows `OmniChip` in tables to request changes in `OmniSearcBar`.
 - Add support for start and end of string support using `^` and `$`.
+- Add support for quoted strings as exact search.
 - Update api-utils to encode URI characters correctly.
+- Add string utils `unquoteString` and `extractQuotedString`
 
 ## 0.7.0 - 2018-10-24